LPS-103504 Rename to basic templates
