+++ modules/apps/layout/layout-admin-web/src/main/java/com/liferay/layout/admin/web/internal/servlet/taglib/clay/SelectBasicTemplatesHorizontalCard.java
@@ -14,17 +14,20 @@
 
 package com.liferay.layout.admin.web.internal.servlet.taglib.clay;
 
-import com.liferay.frontend.taglib.clay.servlet.taglib.soy.VerticalCard;
-import com.liferay.layout.page.template.model.LayoutPageTemplateEntry;
-import com.liferay.portal.kernel.model.LayoutConstants;
+import com.liferay.frontend.taglib.clay.servlet.taglib.soy.HorizontalCard;
+import com.liferay.portal.kernel.language.LanguageUtil;
+import com.liferay.portal.kernel.model.LayoutTypeController;
 import com.liferay.portal.kernel.portlet.LiferayWindowState;
 import com.liferay.portal.kernel.theme.ThemeDisplay;
 import com.liferay.portal.kernel.util.ParamUtil;
 import com.liferay.portal.kernel.util.PortalUtil;
+import com.liferay.portal.kernel.util.ResourceBundleUtil;
 import com.liferay.portal.kernel.util.WebKeys;
+import com.liferay.portal.util.LayoutTypeControllerTracker;
 
 import java.util.HashMap;
 import java.util.Map;
+import java.util.ResourceBundle;
 
 import javax.portlet.PortletURL;
 import javax.portlet.RenderRequest;
@@ -35,15 +38,17 @@ import javax.servlet.http.HttpServletRequest;
 /**
  * @author Eudaldo Alonso
  */
-public class SelectBasicPagesVerticalCard implements VerticalCard {
+public class SelectBasicTemplatesHorizontalCard implements HorizontalCard {
 
-	public SelectBasicPagesVerticalCard(
-		LayoutPageTemplateEntry layoutPageTemplateEntry,
-		RenderRequest renderRequest, RenderResponse renderResponse) {
+	public SelectBasicTemplatesHorizontalCard(
+		String type, RenderRequest renderRequest,
+		RenderResponse renderResponse) {
 
-		_layoutPageTemplateEntry = layoutPageTemplateEntry;
+		_type = type;
 		_renderResponse = renderResponse;
 
+		_layoutTypeController =
+			LayoutTypeControllerTracker.getLayoutTypeController(type);
 		_httpServletRequest = PortalUtil.getHttpServletRequest(renderRequest);
 		_themeDisplay = (ThemeDisplay)renderRequest.getAttribute(
 			WebKeys.THEME_DISPLAY);
@@ -72,11 +77,7 @@ public class SelectBasicPagesVerticalCard implements VerticalCard {
 			addLayoutURL.setParameter(
 				"privateLayout", String.valueOf(privateLayout));
 
-			addLayoutURL.setParameter("type", LayoutConstants.TYPE_CONTENT);
-			addLayoutURL.setParameter(
-				"masterLayoutPageTemplateEntryId",
-				String.valueOf(
-					_layoutPageTemplateEntry.getLayoutPageTemplateEntryId()));
+			addLayoutURL.setParameter("type", _type);
 			addLayoutURL.setWindowState(LiferayWindowState.POP_UP);
 
 			data.put("add-layout-url", addLayoutURL.toString());
@@ -90,7 +91,8 @@ public class SelectBasicPagesVerticalCard implements VerticalCard {
 	@Override
 	public String getElementClasses() {
 		return "add-layout-action-option card-interactive " +
-			"card-interactive-primary";
+			"card-interactive-primary card-type-template " +
+				"template-card-horizontal";
 	}
 
 	@Override
@@ -99,13 +101,14 @@ public class SelectBasicPagesVerticalCard implements VerticalCard {
 	}
 
 	@Override
-	public String getImageSrc() {
-		return _layoutPageTemplateEntry.getImagePreviewURL(_themeDisplay);
-	}
-
-	@Override
 	public String getTitle() {
-		return _layoutPageTemplateEntry.getName();
+		ResourceBundle layoutTypeResourceBundle = ResourceBundleUtil.getBundle(
+			"content.Language", _themeDisplay.getLocale(),
+			_layoutTypeController.getClass());
+
+		return LanguageUtil.get(
+			_httpServletRequest, layoutTypeResourceBundle,
+			"layout.types." + _type);
 	}
 
 	@Override
@@ -114,8 +117,9 @@ public class SelectBasicPagesVerticalCard implements VerticalCard {
 	}
 
 	private final HttpServletRequest _httpServletRequest;
-	private final LayoutPageTemplateEntry _layoutPageTemplateEntry;
+	private final LayoutTypeController _layoutTypeController;
 	private final RenderResponse _renderResponse;
 	private final ThemeDisplay _themeDisplay;
+	private final String _type;
 
 }
