+++ modules/apps/layout/layout-admin-web/src/main/resources/content/Language.properties
@@ -13,7 +13,7 @@ allow-site-administrators-to-modify-this-page-for-their-site-help=Set this to al
 are-you-sure-you-want-to-delete-the-selected-page=Are you sure you want to delete the selected page?
 are-you-sure-you-want-to-reset-your-customizations-to-default=Are you sure you want to reset your customizations to default?
 available-themes=Available Themes
-basic-pages=Basic Pages
+basic-templates=Basic Templates
 by-x=By {0}
 canonical-url=Canonical URL
 change-current-theme=Change Current Theme
